リファクタ：createButtonとclickButtonメソッド以外のビジネスロジックを別ファイルに切り分け
